Add sema web-app dependencies to pypi package CICD pre-commit hooks

+++ submodules/volatility/volatility/renderers/basic.py
@@ -12,21 +12,21 @@ class Bytes(bytes):
 class Address(long):
     """Integer class to allow renderers to differentiate between addresses and numbers"""
     def __new__(cls, number):
-        return long.__new__(cls, number)
+        return int.__new__(cls, number)
 
 
 class Address64(long):
     """Integer class to allow renderers to differentiate between addresses and numbers"""
 
     def __new__(cls, number):
-        return long.__new__(cls, number)
+        return int.__new__(cls, number)
 
 
 class Hex(long):
     """Integer class to allow renderers to differentiate between addresses and numbers"""
 
     def __new__(cls, number):
-        return long.__new__(cls, number)
+        return int.__new__(cls, number)
 
 
 class Renderer(object):
